Import comments attached to blank cells (fixes #10)
This is done in stages:

1. As comments are associated with non-blank cells, remove them from the map.
2. Iterate through the remaining comments, creating a blank cell for each one.

This required row heights to be cached, which in turn required parsing the 'r'
attribute of rows (to get the row number, to get the row height from the cache),
which meant failing on the JMP file at the row rather than the cell level.  The
JMP test file was adapted manually in order to test both the cell- and row-level
failures.