Allow for unmatched comment--cell pairs in cellcount
This changes the way comment-cell pairs are handled.  Comments on blank cells
don't have a matching cell, so one has to be created for them.  The number to be
created must be known in time for initialising the vectors that will go into the
final returned data frame (when the api change is complete and returns all cells
in a workbook in a single data frame).

It breaks the test "gives informative error for a JMP export (row)" in
"test-compatibility.R", because it now checks for cells-without-r-attributes
before checking rows-without-r-attributes.  The tests and test files have been
blended in consequence.