Fix gradientFill inconsistencies
I think I understand this now.  There are two 'stops', one for each colour, at
position 0, and position 0.5 or 1.  If `type=="path"` then left/right/top/bottom
are used, otherwise a direction in degrees is used, where 0 degrees is to the
right, and 90 is upwards.

The next commit will refactor again to make access easier.