Avoid delay on return to R
The return from Rcpp to R was really slow for large spreadsheets, which I
suspect was due to copying or garbage collection.  It was definitely something
to do with the parsed strings, which were stored in a `std::vector<Rcpp:List>`
(one string in each `Rcpp::List`).  By experiment, using another `Rcpp::List` to
wrap the individual ones is much quicker.