Move tidy_xlsx.gif to man/figures
Another attempt to address #88 and render the gif both online and
locally.

It seems that, unless an image in `vignettes` is used by a vignette,
pkgdown won't copy it to `articles`, so it won't be available to the
README.

This still results in an error locally `"Missing images in 'README.md':
'articles/titanic-screenshot.png', 'articles/styles-screenshot.png'`,
even though those files are literally right there.